app_kivymd/requirements.txt
kivy
kivymd
pyyaml
python-dotenv
flask
pyjnius
cython
buildozer